test: Pull base of AnythingLLMMock into a base class
This will get used for other server mocks

https://github.com/DS4SD/ragnardoc/issues/4
Branch: OpenWebUIDuplicate-4
